+++ vclib-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidc/OidcSiopProtocolTest.kt
@@ -1,5 +1,6 @@
 package at.asitplus.wallet.lib.oidc
 
+import at.asitplus.wallet.lib.LibraryInitializer
 import at.asitplus.wallet.lib.agent.CryptoService
 import at.asitplus.wallet.lib.agent.DefaultCryptoService
 import at.asitplus.wallet.lib.agent.Holder
@@ -8,7 +9,9 @@ import at.asitplus.wallet.lib.agent.IssuerAgent
 import at.asitplus.wallet.lib.agent.Verifier
 import at.asitplus.wallet.lib.agent.VerifierAgent
 import at.asitplus.wallet.lib.data.ConstantIndex
+import at.asitplus.wallet.lib.data.CredentialSubject
 import at.asitplus.wallet.lib.oidvci.decodeFromPostBody
+import at.asitplus.wallet.lib.oidvci.decodeFromUrlQuery
 import at.asitplus.wallet.lib.oidvci.encodeToParameters
 import at.asitplus.wallet.lib.oidvci.formUrlEncode
 import com.benasher44.uuid.uuid4
@@ -17,6 +20,8 @@ import io.kotest.matchers.collections.shouldNotBeEmpty
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
 import kotlinx.coroutines.runBlocking
+import kotlinx.serialization.modules.polymorphic
+import kotlinx.serialization.modules.subclass
 
 class OidcSiopProtocolTest : FreeSpec({
 
@@ -31,6 +36,17 @@ class OidcSiopProtocolTest : FreeSpec({
 
     lateinit var holderOidcSiopProtocol: OidcSiopWallet
     lateinit var verifierOidcSiopProtocol: OidcSiopVerifier
+
+    beforeSpec {
+        LibraryInitializer.registerExtensionLibrary(LibraryInitializer.ExtensionLibraryInfo(
+            credentialScheme = TestCredentialScheme,
+            serializersModule = kotlinx.serialization.modules.SerializersModule {
+                polymorphic(CredentialSubject::class) {
+                    subclass(TestCredential::class)
+                }
+            }
+        ))
+    }
 
     beforeEach {
         holderCryptoService = DefaultCryptoService()
@@ -109,7 +125,7 @@ class OidcSiopProtocolTest : FreeSpec({
         val authnRequestUrlParams = authnRequest.encodeToParameters().formUrlEncode()
         println(authnRequestUrlParams)
 
-        val parsedAuthnRequest: AuthenticationRequestParameters = authnRequestUrlParams.decodeFromPostBody()
+        val parsedAuthnRequest: AuthenticationRequestParameters = authnRequestUrlParams.decodeFromUrlQuery()
         val authnResponse = holderOidcSiopProtocol.createAuthnResponseParams(parsedAuthnRequest).getOrThrow()
         val authnResponseParams = authnResponse.encodeToParameters().formUrlEncode()
         println(authnResponseParams)
@@ -120,4 +136,33 @@ class OidcSiopProtocolTest : FreeSpec({
         result.vp.verifiableCredentials.shouldNotBeEmpty()
     }
 
+    "test specific credential" {
+        holder.storeCredentials(
+            IssuerAgent.newDefaultInstance(
+                DefaultCryptoService(),
+                dataProvider = TestCredentialDataProvider(),
+            ).issueCredentialWithTypes(
+                holder.identifier,
+                listOf(TestCredentialScheme.vcType)
+            ).toStoreCredentialInput()
+        )
+
+        val authnRequest = verifierOidcSiopProtocol.createAuthnRequestUrl(
+            walletUrl,
+            relyingPartyUrl,
+            credentialScheme = TestCredentialScheme
+        )
+        println(authnRequest)
+
+        val authnResponse = holderOidcSiopProtocol.createAuthnResponse(authnRequest).getOrThrow()
+        authnResponse.shouldBeInstanceOf<OidcSiopWallet.AuthenticationResponseResult.Redirect>()
+        println(authnResponse)
+
+        val result = verifierOidcSiopProtocol.validateAuthnResponse(authnResponse.url, relyingPartyUrl)
+        result.shouldBeInstanceOf<OidcSiopVerifier.AuthnResponseResult.Success>()
+        result.vp.verifiableCredentials.shouldNotBeEmpty()
+        result.vp.verifiableCredentials.forEach {
+            it.vc.credentialSubject.shouldBeInstanceOf<TestCredential>()
+        }
+    }
 })