+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Release 2.0.1:
  - `vclib-openid`: Remove `OidcSiopProtocol`, replace with `OidcSiopVerifier` and `OidcSiopWallet`, remove `AuthenticationResponse` and `AuthenticationRequest` holder classes
+ - `vclib-openid`: Update implementation of OIDC SIOPv2 to v1.0.12 (2023-01-01), and of OID4VP to draft 18 (2023-04-21). Still missing requesting single claims, selective disclosure, among other parts
 
 Release 2.0.0:
  - Add `AtomicAttribute2023` as a sample for custom credentials